fix bug

+++ colle_tMedecine/colle_tMedecine/ViewModel/PersonnelViewModel.cs
@@ -165,10 +165,15 @@ namespace colle_tMedecine.ViewModel
         {
             ServiceUser.ServiceUserClient service = new ServiceUser.ServiceUserClient();
             Model.User user = (Model.User)param;
-            this._allUser.Remove(user);
-            if (service.DeleteUser(user.Login))
+             View.MainWindow mainwindow = (View.MainWindow)Application.Current.MainWindow;
+            ViewModel.MainWindow mainwindowVM = (ViewModel.MainWindow)mainwindow.DataContext;
+            if (user.Login !=  mainwindowVM.ConnectedUser.Login)
             {
-                ListUser = new ObservableCollection<Model.User>(this._allUser);
+                this._allUser.Remove(user);
+                if (service.DeleteUser(user.Login))
+                {
+                    ListUser = new ObservableCollection<Model.User>(this._allUser);
+                }
             }
 
         }
